ENGR00137894-5 MX53 ARD:Support splashimage
This patch supports to use pwm wave to control
backlight. The pwm rate is 200Hz and the pwm
duty is 50%. Use 'lvds_num' env variable to
choose to use lvds0 or lvds1. Note that you
need to add 'splashimage' env variable to
set the memory address of the bmp image.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>