ENGR00219636 MX6 UBOOT : Enlarge command line buffer size
There is a problem that a too long command line parameter in U-Boot
console will actually be truncated and not properly truncated to the
kernel.

The root cause is that the command line in the U-Boot console is read
into a buffer --- console_buffer[CONFIG_SYS_CBSIZE]. Currently the
CONFIG_SYS_CBSIZE is set as 256. Command line parameter larger than it
will not be recorded. On the other hand, max length of boot parameter of
linux kernel is set to 1024, which means it can accept parameter size as
large as 1024.

So we need to align these 2 values. Enlarge CONFIG_SYS_CBSIZE to 1024 as
well.

Signed-off-by: Eric Sun <jian.sun@freescale.com>